update : clean repo add schema to every actions
