+++ src/lib/agent/plugins/core/account/actions/signMessage.ts
