+++ src/lib/agent/plugins/core/account/actions/verifyMessage.ts
