+++ src/lib/agent/plugins/core/contract/actions/declareContract.ts
