+++ src/lib/agent/plugins/core/rpc/actions/getBlockLatestAccepted.ts
