+++ src/lib/agent/plugins/core/rpc/actions/getBlockNumber.ts
