+++ src/lib/agent/plugins/core/rpc/actions/getBlockWithTxs.ts
