+++ src/lib/agent/plugins/core/rpc/actions/getChainId.ts
