+++ src/lib/agent/plugins/core/rpc/actions/getSpecVersion.ts
