+++ src/lib/agent/plugins/core/rpc/actions/getTransactionByBlockIdAndIndex.ts
