+++ src/lib/agent/plugins/unruggable/actions/getLockedLiquidity.ts
@@ -1,4 +1,4 @@
-import { ContractAddressParams } from 'src/lib/agent/schemas/schema';
+import { ContractAddressParams } from 'src/lib/agent/plugins/unruggable/schema';
 import { Contract } from 'starknet';
 import { FACTORY_ABI } from '../abis/unruggableFactory';
 import { FACTORY_ADDRESS } from '../constants';
