+++ test/unit-test/method/account/estimateAccountDeployFee.spec.ts
@@ -1,4 +1,4 @@
-import { estimateAccountDeployFee } from 'src/lib/agent/plugins/core/account/estimateAccountDeployFee';
+import { estimateAccountDeployFee } from 'src/lib/agent/plugins/core/account/actions/estimateAccountDeployFee';
 import * as C from '../../../utils/constant';
 
 describe('Estimate Account Deploy Fee', () => {
